my-azure-auth-app/requirement.txt
Flask==2.0.1
firebase-admin==5.0.0
pybcrypt==0.4